pyproject.toml
[tool.poetry]
name = "telebot"
version = "0.1.0"
description = ""
authors = ["fikrehm <fikrehmariam77@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytelegrambotapi = "^4.23.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
